installing the rsdis

+++ controller/api/v1/posts_api.js
@@ -20,8 +20,8 @@ module.exports.destroy= async function(req,res)
     try{ 
    let post=await Post.findById(req.params.id);
         // .id means converting object id into string
-      ///  if(post.user==req.user.id)
-       // {
+       if(post.user==req.user.id)
+       {
             post.remove();
            await Comment.deleteMany({post:req.params.id});
            
@@ -29,11 +29,13 @@ module.exports.destroy= async function(req,res)
                 return res.json(200,{
                     message:"Post and associated comments deleted succesfully !"
                 });
-       // }
-       /// else{
-         ///   req.flash('error','You can not delete this post');
-          //  return res.redirect('back');
-      //  }
+       }
+       else{
+         //  req.flash('error','You can not delete this post');
+           return res.json(401,{
+               message:"You cannot delete this post !"
+           })
+      }
     }catch(err){
         console.log('******',err);
         return res.json(500,{
